create infrastructure and dbcontext ,exception,validators

+++ HR_Management.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using HR_Management.Application.Exceptions;
 using HR_Management.Application.Features.LeaveRequests.Requests.Commands;
 using HR_Management.Application.Persistence.Contracts;
 using MediatR;
@@ -19,6 +20,11 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
     public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
     {
         var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+        if (leaveRequest == null)
+        {
+            throw new NotFoundException(nameof(leaveRequest), request.Id);
+        }
+
         await _leaveRequestRepository.Delete(leaveRequest);
         return Unit.Value;
     }
