+++ HR_Management.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using HR_Management.Application.Exceptions;
 using HR_Management.Application.Features.LeaveTypes.Requests.Commands;
 using HR_Management.Application.Persistence.Contracts;
 using MediatR;
@@ -19,6 +20,10 @@ public class DeleteLeaveTypeCommandHandler:IRequestHandler<DeleteLeaveTypeComman
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveType = await _leaveTypeRepository.Get(request.Id);
+        if (leaveType == null)
+        {
+            throw new NotFoundException(nameof(leaveType), request.Id);
+        }
         await _leaveTypeRepository.Delete(leaveType);
         return Unit.Value;
     }
